Add precondition validation
What: The request payload is validated before attempting to persist it. Defining a validation method in the BeersRequest structure, which represents the HTTP request to create or update a beer, located in the beer.go file of the model package.

Why: To prevent, for example, errors from being thrown in the database when assigning null values to non-nullable fields.

DROP-beer-api