Change our beer handler
What: The `beer_handler.go` script is modified so that it can implement the constructor and “injection” of beer_service.go. Delegating business logic to the corresponding service. In addition to handling responses and errors.

Why: The handler must be allowed to receive requests, validate them, and delegate business logic to the corresponding service. In addition to handling responses and errors.

DROP-beer-api