Add Swagger to the project
What: `Swagger` is added to the project, which is a tool that allows us to self-document REST APIs.

Why: So that an additional JSON schema can be included that describes the structure of the request body sent to an operation in a REST API and the structure of the response bodies returned from an operation.

DROP-beer-api